CNV-35729: Improve modal save while there are changes occurred in the back-end - part 2 network
