+++ src/views/virtualmachines/details/tabs/configuration/network/components/modal/VirtualMachinesEditNetworkInterfaceModal.tsx
@@ -1,6 +1,5 @@
 import React, { FC, useCallback } from 'react';
 
-import VirtualMachineModel from '@kubevirt-ui/kubevirt-api/console/models/VirtualMachineModel';
 import { V1Interface, V1Network, V1VirtualMachine } from '@kubevirt-ui/kubevirt-api/kubevirt';
 import NetworkInterfaceModal from '@kubevirt-utils/components/NetworkInterfaceModal/NetworkInterfaceModal';
 import {
@@ -11,7 +10,6 @@ import {
 import { useKubevirtTranslation } from '@kubevirt-utils/hooks/useKubevirtTranslation';
 import { getInterfaces, getNetworks } from '@kubevirt-utils/resources/vm';
 import { NetworkPresentation } from '@kubevirt-utils/resources/vm/utils/network/constants';
-import { k8sUpdate } from '@openshift-console/dynamic-plugin-sdk';
 
 type VirtualMachinesEditNetworkInterfaceModalProps = {
   isOpen: boolean;
@@ -45,14 +43,8 @@ const VirtualMachinesEditNetworkInterfaceModal: FC<
             []),
           resultInterface,
         ];
-        const updatedVM = updateVMNetworkInterfaces(vm, updatedNetworks, updatedInterfaces);
 
-        return k8sUpdate({
-          data: updatedVM,
-          model: VirtualMachineModel,
-          name: updatedVM.metadata.name,
-          ns: updatedVM.metadata.namespace,
-        });
+        return updateVMNetworkInterfaces(vm, updatedNetworks, updatedInterfaces);
       },
     [vm, nicPresentation],
   );
